pdaj/table/serializers.py
from rest_framework import serializers
from .models import Table


class TableSerializer(serializers.ModelSerializer):
    class Meta:
        model = Table
        fields = ['result','time_in_sec','max_memory_in_MB']