pyproject.toml
[tool.poetry]
name = "pdaj"
version = "0.1.0"
description = ""
authors = ["vuk.saric <vuk.saric12@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
markdown = "^3.*"
django-filter="^21.*"
psycopg2-binary= "^2.9.*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
